CPR-637 move canonical API into a different controller
