+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/api/controller/SearchController.kt
@@ -13,35 +13,19 @@ import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RestController
 import uk.gov.justice.digital.hmpps.personrecord.api.constants.Roles
-import uk.gov.justice.digital.hmpps.personrecord.api.controller.exceptions.PersonKeyNotFoundException
 import uk.gov.justice.digital.hmpps.personrecord.api.controller.exceptions.PersonRecordNotFoundException
 import uk.gov.justice.digital.hmpps.personrecord.api.model.PersonIdentifierRecord
-import uk.gov.justice.digital.hmpps.personrecord.api.model.canonical.CanonicalRecord
 import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.PersonEntity
 import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.PersonEntity.Companion.extractSourceSystemId
-import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.PersonKeyEntity
-import uk.gov.justice.digital.hmpps.personrecord.jpa.repository.PersonKeyRepository
 import uk.gov.justice.digital.hmpps.personrecord.jpa.repository.PersonRepository
 import uk.gov.justice.hmpps.kotlin.common.ErrorResponse
-import java.util.UUID
 
 @Tag(name = "Search")
 @RestController
 @PreAuthorize("hasRole('${Roles.SEARCH_API_READ_ONLY}')")
 class SearchController(
   private val personRepository: PersonRepository,
-  private val personKeyRepository: PersonKeyRepository,
-
 ) {
-
-  @Operation(description = "Retrieve person record by UUID")
-  @GetMapping("/search/person/{uuid}")
-  @ApiResponses(
-    ApiResponse(responseCode = "200", description = "OK"),
-  )
-  fun getCanonicalRecord(
-    @PathVariable(name = "uuid") uuid: UUID,
-  ): CanonicalRecord = buildCanonicalRecord(personKeyRepository.findByPersonId(uuid), uuid)
 
   @Operation(description = "Search for person record and associated records with a CRN within the system")
   @GetMapping("/search/offender/{crn}")
@@ -112,11 +96,6 @@ class SearchController(
   private fun handlePersonRecord(personEntity: PersonEntity?, identifier: String): List<PersonIdentifierRecord> = when {
     personEntity != PersonEntity.empty -> buildListOfLinkedRecords(personEntity)
     else -> throw PersonRecordNotFoundException(identifier)
-  }
-
-  private fun buildCanonicalRecord(personKeyEntity: PersonKeyEntity?, uuid: UUID): CanonicalRecord = when {
-    personKeyEntity?.personEntities?.isNotEmpty() == true -> CanonicalRecord.from(personKeyEntity)
-    else -> throw PersonKeyNotFoundException(uuid)
   }
 
   private fun buildListOfLinkedRecords(personEntity: PersonEntity): List<PersonIdentifierRecord> = personEntity.personKey?.personEntities?.mapNotNull {
